+++ pmd-xml/src/main/java/net/sourceforge/pmd/lang/xml/XmlParser.java
@@ -13,6 +13,7 @@ import net.sourceforge.pmd.lang.ParserOptions;
 import net.sourceforge.pmd.lang.TokenManager;
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.ast.ParseException;
+import net.sourceforge.pmd.lang.xml.ast.internal.XmlParserImpl;
 
 /**
  * Adapter for the XmlParser.
@@ -35,7 +36,7 @@ public class XmlParser extends AbstractParser {
 
     @Override
     public Node parse(String fileName, Reader source) throws ParseException {
-        return new net.sourceforge.pmd.lang.xml.ast.XmlParser((XmlParserOptions) parserOptions).parse(source);
+        return new XmlParserImpl((XmlParserOptions) parserOptions).parse(source);
     }
 
     @Override
